LayoutTests/fast/forms/vertical-writing-mode/select-menulist-dimensions-expected.txt

Test that vertical writing mode selects have swapped dimensions compared to a horizontal one

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


vertical-rl
PASS verticalRL.offsetWidth is horizontal.offsetHeight
PASS verticalRL.offsetHeight is horizontal.offsetWidth

vertical-lr
PASS verticalLR.offsetWidth is horizontal.offsetHeight
PASS verticalLR.offsetHeight is horizontal.offsetWidth

sideways-rl
PASS sidewaysRL.offsetWidth is horizontal.offsetHeight
PASS sidewaysRL.offsetHeight is horizontal.offsetWidth

sideways-lr
PASS sidewaysLR.offsetWidth is horizontal.offsetHeight
PASS sidewaysLR.offsetHeight is horizontal.offsetWidth

direction: rtl should not change dimensions

horizontal rtl
PASS horizontalRTL.offsetWidth is horizontal.offsetWidth
PASS horizontalRTL.offsetHeight is horizontal.offsetHeight

vertical-rl rtl
PASS verticalRL_RTL.offsetWidth is verticalRL.offsetWidth
PASS verticalRL_RTL.offsetHeight is verticalRL.offsetHeight

vertical-lr rtl
PASS verticalLR_RTL.offsetWidth is verticalLR.offsetWidth
PASS verticalLR_RTL.offsetHeight is verticalLR.offsetHeight

sideways-rl rtl
PASS sidewaysRL_RTL.offsetWidth is sidewaysRL.offsetWidth
PASS sidewaysRL_RTL.offsetHeight is sidewaysRL.offsetHeight

sideways-lr rtl
PASS sidewaysLR_RTL.offsetWidth is sidewaysLR.offsetWidth
PASS sidewaysLR_RTL.offsetHeight is sidewaysLR.offsetHeight
PASS successfullyParsed is true

TEST COMPLETE
